Refactor(cli): Split metrics, errors and error context
metrics.py (669 lines), error_context.py (592) and errors.py (470)
all sat past the 400-line reviewability limit the aislop code-quality
gate applies.

Split each into a package at the same dotted path: errors becomes
base, types and messages; error_context becomes models, the config
and network detector groups, and the auto-detect dispatcher; metrics
becomes models, context, collector and session.

The metrics singleton and its accessors live in session alone, so
there is still exactly one global collector; the conftest autouse
fixture resets the same instance it did before.
