+++ src/cli/errors/types.py
@@ -1,113 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
 
-"""
-CLI Error Classes
-
-Enhanced error classes for better user experience with actionable suggestions
-and documentation links.
-
-Phase 9: CLI & UX Improvements
-Phase 13, Step 4: Enhanced Error Messages with Context
-"""
+"""Concrete CLI error classes for configuration, arguments, API and I/O."""
 
 from typing import Any
 
-
-class CLIError(Exception):
-    """
-    Base CLI error with helpful context.
-
-    This error class provides structured error information including:
-    - Clear error message
-    - Actionable suggestions for resolution
-    - Links to relevant documentation
-
-    Example:
-        >>> raise CLIError(
-        ...     "Configuration file not found: config.yaml",
-        ...     suggestion="Create a config.yaml file or specify path with --config",
-        ...     doc_link="https://docs.example.com/configuration"
-        ... )
-    """
-
-    def __init__(
-        self,
-        message: str,
-        suggestion: str | None = None,
-        doc_link: str | None = None,
-        context: dict[str, Any] | None = None,
-        recovery_hints: list[str] | None = None,
-    ):
-        """
-        Initialize CLI error.
-
-        Args:
-            message: Error message describing what went wrong
-            suggestion: Optional suggestion for how to fix the error
-            doc_link: Optional link to relevant documentation
-            context: Optional context information dictionary
-            recovery_hints: Optional list of step-by-step recovery instructions
-        """
-        self.message = message
-        self.suggestion = suggestion
-        self.doc_link = doc_link
-        self.context = context or {}
-        self.recovery_hints = recovery_hints or []
-        super().__init__(message)
-
-    def __str__(self) -> str:
-        """Format error with suggestions and documentation."""
-        parts = [f"❌ Error: {self.message}"]
-
-        # Context information
-        if self.context:
-            parts.append("\n📋 Context:")
-            for key, value in self.context.items():
-                parts.append(f"  • {key}: {value}")
-
-        # Recovery hints (step-by-step)
-        if self.recovery_hints:
-            parts.append("\n🔧 How to fix:")
-            for i, hint in enumerate(self.recovery_hints, 1):
-                parts.append(f"  {i}. {hint}")
-
-        # Simple suggestion (backward compatible)
-        elif self.suggestion:
-            parts.append(f"\n💡 Suggestion: {self.suggestion}")
-
-        # Documentation link
-        if self.doc_link:
-            parts.append(f"\n📖 Documentation: {self.doc_link}")
-
-        return "\n".join(parts)
-
-    def add_context(self, key: str, value: Any) -> "CLIError":
-        """
-        Add context information to error.
-
-        Args:
-            key: Context key
-            value: Context value
-
-        Returns:
-            Self for chaining
-        """
-        self.context[key] = value
-        return self
-
-    def add_recovery_hint(self, hint: str) -> "CLIError":
-        """
-        Add a recovery hint.
-
-        Args:
-            hint: Recovery instruction
-
-        Returns:
-            Self for chaining
-        """
-        self.recovery_hints.append(hint)
-        return self
+from .base import CLIError
 
 
 class ConfigurationError(CLIError):
@@ -374,97 +272,3 @@ class NetworkError(CLIError):
             context=context,
             recovery_hints=recovery_hints or default_hints,
         )
-
-
-def format_validation_errors(errors: list[dict[str, Any]]) -> str:
-    """
-    Format multiple validation errors into readable message.
-
-    Args:
-        errors: List of validation error dictionaries with 'path' and 'message'
-
-    Returns:
-        Formatted error message
-
-    Example:
-        >>> errors = [
-        ...     {'path': 'project', 'message': 'Required field missing'},
-        ...     {'path': 'api.github.token', 'message': 'Invalid token format'}
-        ... ]
-        >>> print(format_validation_errors(errors))
-        Configuration validation failed with 2 error(s):
-          - project: Required field missing
-          - api.github.token: Invalid token format
-    """
-    if not errors:
-        return "Validation passed"
-
-    lines = [f"Configuration validation failed with {len(errors)} error(s):"]
-    for error in errors:
-        path = error.get("path", "unknown")
-        message = error.get("message", "Unknown error")
-        lines.append(f"  - {path}: {message}")
-
-    return "\n".join(lines)
-
-
-def suggest_common_fixes(error: Exception) -> str | None:
-    """
-    Suggest common fixes based on error type and message.
-
-    Args:
-        error: The exception that was raised
-
-    Returns:
-        Suggestion string or None if no common fix available
-
-    Example:
-        >>> error = FileNotFoundError("config.yaml")
-        >>> print(suggest_common_fixes(error))
-        Create a config.yaml file using config.example.yaml as template
-    """
-    error_str = str(error).lower()
-
-    # Check error type first
-    if isinstance(error, FileNotFoundError):
-        if "config" in error_str:
-            return "Create a config.yaml file using config.example.yaml as template"
-        return "Verify the file path exists and is accessible"
-
-    # Common error patterns and suggestions
-    if "config.yaml" in error_str and "not found" in error_str:
-        return "Create a config.yaml file using config.example.yaml as template"
-
-    if "permission denied" in error_str:
-        return "Check file permissions or run with appropriate privileges"
-
-    if "connection" in error_str or "network" in error_str:
-        return "Check network connectivity and firewall settings"
-
-    if "authentication" in error_str or "401" in error_str:
-        return "Verify API credentials and tokens are correct"
-
-    if "not found" in error_str and "repository" in error_str:
-        return "Verify repository path exists and is accessible"
-
-    if "disk" in error_str or "space" in error_str:
-        return "Free up disk space or use --no-cache flag"
-
-    if "yaml" in error_str and ("invalid" in error_str or "parse" in error_str):
-        return "Check YAML syntax - ensure proper indentation and no tabs"
-
-    return None
-
-
-__all__ = [
-    "CLIError",
-    "ConfigurationError",
-    "InvalidArgumentError",
-    "APIError",
-    "PermissionError",
-    "DiskSpaceError",
-    "ValidationError",
-    "NetworkError",
-    "format_validation_errors",
-    "suggest_common_fixes",
-]
