+++ src/cli/metrics/collector.py
@@ -1,24 +1,7 @@
-#!/usr/bin/env python3
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
 
-"""
-Performance Metrics Module
-
-Collects and reports performance metrics for repository analysis operations.
-Provides execution summaries, timing breakdowns, resource usage statistics,
-and debug-level profiling information.
-
-Features:
-- Execution timing and breakdown
-- Resource usage tracking (memory, CPU, disk I/O)
-- API call statistics
-- Operation profiling
-- Debug mode detailed metrics
-- Beautiful formatted output
-
-Phase 13: CLI & UX Improvements - Step 6
-"""
+"""Thread-safe collector for timings, API statistics and resource usage."""
 
 # This module renders user-facing terminal output (performance summaries and
 # debug metrics); print() is the intended sink here, not leftover debugging.
@@ -28,156 +11,24 @@ import logging
 import threading
 import time
 from collections import defaultdict
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
 
 import psutil
 
+from .context import _TimingContext
+from .models import (
+    APIStatistics,
+    OperationMetrics,
+    ResourceUsage,
+    TimingMetric,
+    format_bytes,
+    format_duration,
+    format_percentage,
+)
+
 
 logger = logging.getLogger(__name__)
-
-
-# =============================================================================
-# DATA STRUCTURES
-# =============================================================================
-
-
-@dataclass
-class TimingMetric:
-    """Individual timing measurement."""
-
-    name: str
-    duration: float
-    start_time: float
-    end_time: float
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def __str__(self) -> str:
-        """Format timing for display."""
-        return f"{self.name}: {format_duration(self.duration)}"
-
-
-@dataclass
-class APIStatistics:
-    """API call statistics."""
-
-    api_name: str
-    total_calls: int = 0
-    cached_calls: int = 0
-    failed_calls: int = 0
-    total_duration: float = 0.0
-
-    @property
-    def cache_hit_rate(self) -> float:
-        """Calculate cache hit rate."""
-        if self.total_calls == 0:
-            return 0.0
-        return (self.cached_calls / self.total_calls) * 100
-
-    @property
-    def average_duration(self) -> float:
-        """Calculate average call duration."""
-        if self.total_calls == 0:
-            return 0.0
-        return self.total_duration / self.total_calls
-
-    @property
-    def calls_per_second(self) -> float:
-        """Calculate calls per second."""
-        if self.total_duration == 0:
-            return 0.0
-        return self.total_calls / self.total_duration
-
-
-@dataclass
-class ResourceUsage:
-    """Resource usage statistics."""
-
-    peak_memory_mb: float = 0.0
-    avg_memory_mb: float = 0.0
-    cpu_time_seconds: float = 0.0
-    cpu_utilization: float = 0.0
-    disk_read_mb: float = 0.0
-    disk_write_mb: float = 0.0
-
-
-@dataclass
-class OperationMetrics:
-    """Metrics for a specific operation (e.g., repository analysis)."""
-
-    operation_name: str
-    duration: float
-    success: bool = True
-    error: str | None = None
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-# =============================================================================
-# FORMATTING HELPERS
-# =============================================================================
-
-
-def format_duration(seconds: float) -> str:
-    """
-    Format duration in human-readable format.
-
-    Args:
-        seconds: Duration in seconds
-
-    Returns:
-        Formatted string (e.g., "2m 15s", "1h 23m 45s")
-    """
-    if seconds < 1:
-        return f"{seconds * 1000:.0f}ms"
-
-    if seconds < 60:
-        return f"{seconds:.1f}s"
-
-    minutes = int(seconds // 60)
-    secs = int(seconds % 60)
-
-    if minutes < 60:
-        return f"{minutes}m {secs}s"
-
-    hours = minutes // 60
-    minutes = minutes % 60
-    return f"{hours}h {minutes}m {secs}s"
-
-
-def format_bytes(bytes_count: float) -> str:
-    """
-    Format bytes in human-readable format.
-
-    Args:
-        bytes_count: Number of bytes
-
-    Returns:
-        Formatted string (e.g., "1.2 GB", "345 MB")
-    """
-    for unit in ["B", "KB", "MB", "GB", "TB"]:
-        if bytes_count < 1024.0:
-            return f"{bytes_count:.1f} {unit}"
-        bytes_count /= 1024.0
-    return f"{bytes_count:.1f} PB"
-
-
-def format_percentage(value: float, total: float) -> str:
-    """
-    Format percentage with value.
-
-    Args:
-        value: Partial value
-        total: Total value
-
-    Returns:
-        Formatted string (e.g., "45s (33%)")
-    """
-    if total == 0:
-        return f"{format_duration(value)} (0%)"
-
-    percentage = (value / total) * 100
-    return f"{format_duration(value)} ({percentage:.0f}%)"
 
 
 # =============================================================================
@@ -541,129 +392,3 @@ class MetricsCollector:
                 lines.append(f"  {format_name}: {file_path} (not created)")
 
         return "\n".join(lines)
-
-
-class _TimingContext:
-    """Context manager for timing operations."""
-
-    def __init__(self, collector: MetricsCollector, name: str, metadata: dict[str, Any]):
-        """Initialize timing context."""
-        self.collector = collector
-        self.name = name
-        self.metadata = metadata
-        self.start_time = 0.0
-
-    def __enter__(self):
-        """Start timing."""
-        self.start_time = time.time()
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """End timing and record."""
-        end_time = time.time()
-        duration = end_time - self.start_time
-
-        self.collector.record_timing(
-            self.name, duration, self.start_time, end_time, **self.metadata
-        )
-
-        return False
-
-
-# =============================================================================
-# GLOBAL INSTANCE
-# =============================================================================
-
-# Global metrics collector instance
-_global_collector: MetricsCollector | None = None
-
-
-def get_metrics_collector() -> MetricsCollector:
-    """
-    Get global metrics collector instance.
-
-    Returns:
-        MetricsCollector instance
-    """
-    global _global_collector
-    if _global_collector is None:
-        _global_collector = MetricsCollector()
-    return _global_collector
-
-
-def reset_metrics_collector():
-    """Reset global metrics collector."""
-    global _global_collector
-    if _global_collector:
-        _global_collector.finalize()
-    _global_collector = MetricsCollector()
-
-
-# =============================================================================
-# CONVENIENCE FUNCTIONS
-# =============================================================================
-
-
-def time_operation(name: str, **metadata):
-    """
-    Time an operation using global collector.
-
-    Args:
-        name: Operation name
-        **metadata: Additional metadata
-
-    Returns:
-        Context manager
-    """
-    return get_metrics_collector().time_operation(name, **metadata)
-
-
-def record_api_call(api_name: str, duration: float, cached: bool = False, failed: bool = False):
-    """
-    Record an API call using global collector.
-
-    Args:
-        api_name: Name of the API
-        duration: Call duration
-        cached: Whether cached
-        failed: Whether failed
-    """
-    get_metrics_collector().record_api_call(api_name, duration, cached, failed)
-
-
-def print_performance_summary(verbose: bool = False):
-    """
-    Print performance summary using global collector.
-
-    Args:
-        verbose: Include detailed breakdown
-    """
-    get_metrics_collector().print_summary(verbose)
-
-
-def print_debug_metrics():
-    """Print debug metrics using global collector."""
-    get_metrics_collector().print_debug_metrics()
-
-
-__all__ = [
-    # Data structures
-    "TimingMetric",
-    "APIStatistics",
-    "ResourceUsage",
-    "OperationMetrics",
-    # Main class
-    "MetricsCollector",
-    # Global instance
-    "get_metrics_collector",
-    "reset_metrics_collector",
-    # Convenience functions
-    "time_operation",
-    "record_api_call",
-    "print_performance_summary",
-    "print_debug_metrics",
-    # Formatting
-    "format_duration",
-    "format_bytes",
-    "format_percentage",
-]
